Added the Command-T plugin for fuzzy file search.
I added this because it seems to be faster or used more than
FuzzyFinder. It does have a few added complexities though. It
requires a vim that has ruby support compiled in so I had to build
ruby from source on my Mac. It also has a paired C component that
needs to be compiled for it to work. These added complexities give
the wanted speed so I put up with them.