Added autocmd to mkdir -p before saving files.
I added a snippet I obtained from stack overflow that makes it so that
before saving it checks if the path of the file exists and if it doesn't
exist then it simply mkdir -p the directory so that vim can properly
save the file. I added this because I found it annoying to have to
switch to the command line to create new files in a specific path. I
also found it annoying to have to run the shell commands from vim.