Made RunTest work for cucumber tests.
There was an issue where RunTests with a given filename wasn't working
properly for cucumber tests because it was checking for the .feature
extension at the end of the filename argument. However, the filename
argument isn't really a filename it is the arguments for the command
hence I needed to modify the regex that determines if it is a cucumber
test or not that is being run.