Got rid of full backtrace in individual test run.
I did this because the backtraces are so long that I would then
have to switch to the mouse and scroll to the top to find anything of
value in the backtrace. That eventually got annoying enough that I went
in and removed the -b. We will see if I need to add it back or come up
with a better solution somehow.