Set the foldlevel to 1.
I did this so that when I open a file by default it automatically folds
everything at a fold level depth of 1. 0 is the original default and that
folds everything. The larger the number the less gets auto folded because the
deeper the folding depth before it starts auto folding.