Disabled cursorcolumn and cursorline.
I did this for performance reasons because it was making horizontal movement
in the terminal version of Vim rediculously slow when in ruby files with
syntax highlighting on. Therefore, I changed it to be off by default and
implement <leader>p to PingCursor which shows the cursorline and column
briefly so that my eyes can track it.