Added the vim-ruby-refactoring plugin.
I did this just to play around with it and see if I would actually use any of
the functionality it provides in my day to day usage. After briefly playing
around with it the only ones I think I might actually use are the RSpec let
one, and the add parameter to function. Not sure if this is really worth it
just for those two, but we will see.