Added autocmds to set foldmethod=manual for insert.
I did this beacuse when foldmethod=syntax is set for insert mode and you go to
do wordcompletion it totally tanks performance for some reason. So, if you
just set it to manual when you enter insert mode and then set it back to what
it was configured as before when you leave insert, all works as expected with
decent performance.