Switched to Selecta for Fuzzy-Matching.
I did this because it is Gary Bernhardts latest bad ass unix tool. It is a
generic unix command line fuzzy matcher that doesn't care what it fuzzy
matches. So you can pipe whatever you want into it and fuzzy match. It is
focused largely on being extremely fast and targeted at doing what it does
very well. I have ported over all of my fuzzy matching mappings to use
Selecta. It has no caching which is ideal if it is fast enough.