Made <leader>f exclude tags file & cur edit buffer.
I did this because I never want to open the tags file as it is used directly
by vim via other mechanisms. Also, I modified it to exclude the buffer that is
currently being edited. I also found out that the find command I had before
wasn't working 100% correctly and excluding everything it should because I was
missing the -print action at the very end of it.