Added SnipMate, Supertab, AutoPair.
I added SnipMate so that I could build up a snippet library that would
eliminate some of the reptitious code generation. I added Supertab so that I
could use Tab for code completion in addition to the SnipMate trigger. I added
AutoPair so that when I insert quotes, parens, braces, etc. it does nice
things and inserts the second of the pair and handles deleting of pairs etc.