Package selecta in this repo.
I did this so that it doesn't have to be worried about as an external
dependency. Instead it is just included as part of this repo and the vimrc
references the selecta included in this repo. It also lets us control upgrades
of it so that we know they are properly working versions.