Document opening ruby file edit performance problem.
I did this because I was having weird performance problems simply opening and
editing a ruby file and having it a take a few when it should have been
instant. I did some digging and testing of all the plugins and options and
narrowed it down to the documented features in here causing it.