Upgrade vim-test-recall plugin
Why you made the change:

I did this so that help tag generation wouldn't cause problems with Git.

How the change addresses the need:

Upgrading fixes these problems because the latest version has added
doc/tags to the .gitignore of vim-test-recall so that it being generated
doesn't cause a problem anymore.