Add & configure vim-multicursor plugin
Why you made the change:

I did this so that I could have multiple cursor support that I would be
happy with. I had vim-multiple-cursors plugin before but I wasn't happy
with it for a number of reasons. Therefore, I figured might as well try
this one and I like it at least initially relatively well. We will see
after I use it a bit longer. But, I at least like it enough to add it
officially to my dotvim.